It shows text for game over or you win cases
