+++ app/src/main/java/com/example/two_zero_four_eight/di/UseCasesModule.kt
@@ -1,7 +1,9 @@
 package com.example.two_zero_four_eight.di
 
 import com.example.two_zero_four_eight.use_cases.AddNumberToBoardGameUseCase
+import com.example.two_zero_four_eight.use_cases.CombineAndMoveNumbersUseCase
 import com.example.two_zero_four_eight.use_cases.CreateBoardGameUseCase
+import com.example.two_zero_four_eight.use_cases.HasWonTheGameUseCase
 import com.example.two_zero_four_eight.use_cases.IsTherePossibleMovesUseCase
 import com.example.two_zero_four_eight.use_cases.MoveNumbersUseCase
 import dagger.Module
@@ -14,29 +16,34 @@ import dagger.hilt.components.SingletonComponent
 object UseCasesModule {
 
     @Provides
-    fun provideAddNumberToBoardGameUseCase(
-    ): AddNumberToBoardGameUseCase {
-        return AddNumberToBoardGameUseCase()
-    }
+    fun provideCombineAndMoveNumberUseCase() =
+        CombineAndMoveNumbersUseCase()
+
+    @Provides
+    fun provideAddNumberToBoardGameUseCase() =
+        AddNumberToBoardGameUseCase()
 
     @Provides
     fun provideCreateBoardGameUseCase(
         addNumberUseCase: AddNumberToBoardGameUseCase
-    ): CreateBoardGameUseCase {
-        return CreateBoardGameUseCase(addNumberUseCase)
-    }
+    ) =
+        CreateBoardGameUseCase(addNumberUseCase)
 
     @Provides
-    fun provideIsTherePossibleMovesUseCase(
-    ): IsTherePossibleMovesUseCase {
-        return IsTherePossibleMovesUseCase()
-    }
+    fun provideIsTherePossibleMovesUseCase() =
+        IsTherePossibleMovesUseCase()
+
+    @Provides
+    fun provideHasWonTheGameUseCase() =
+        HasWonTheGameUseCase()
 
     @Provides
     fun provideMoveNumbersUseCase(
-        useCase1: AddNumberToBoardGameUseCase,
-        useCase2: IsTherePossibleMovesUseCase
+        useCase1: CombineAndMoveNumbersUseCase,
+        useCase2: AddNumberToBoardGameUseCase,
+        useCase3: IsTherePossibleMovesUseCase,
+        useCase4: HasWonTheGameUseCase
     ): MoveNumbersUseCase {
-        return MoveNumbersUseCase(useCase1, useCase2)
+        return MoveNumbersUseCase(useCase1, useCase2, useCase3, useCase4)
     }
 }
