+++ app/src/main/java/com/example/two_zero_four_eight/model/GameState.kt
@@ -1,6 +1,16 @@
 package com.example.two_zero_four_eight.model
 
+import com.example.two_zero_four_eight.model.GameStatus.*
+import com.example.two_zero_four_eight.ui.DEFAULT_NUMBER_TO_WIN
+
 data class GameState(
     val board: MutableList<MutableList<Int>> = mutableListOf(),
-    var isGameOver: Boolean = false,
+    var gameStatus: GameStatus = PLAYING,
+    var numberToWin: Int = DEFAULT_NUMBER_TO_WIN
 )
+
+enum class GameStatus {
+    PLAYING,
+    GAME_OVER,
+    YOU_WIN
+}
