+++ app/src/main/java/com/example/two_zero_four_eight/use_cases/IsTherePossibleMovesUseCase.kt
@@ -1,5 +1,6 @@
 package com.example.two_zero_four_eight.use_cases
 
+import com.example.two_zero_four_eight.model.GameStatus.*
 import com.example.two_zero_four_eight.use_cases.utils.MoveNumberResult
 
 class IsTherePossibleMovesUseCase {
@@ -12,7 +13,11 @@ class IsTherePossibleMovesUseCase {
      * vertically or horizontally with the same number), keep playing.
      * 3) Otherwise is game over
      * **/
-    fun existMovesToContinue(boardGame: MutableList<MutableList<Int>>): MoveNumberResult {
+    fun checkMovesToContinue(
+        boardGame: MutableList<MutableList<Int>>,
+        numberToWin: Int
+    ): MoveNumberResult {
+        val playingBoardGame = MoveNumberResult(boardGame, numberToWin =  numberToWin)
         val boardSize = boardGame.size
 
         for (rowIndex in 0..<boardSize) {
@@ -20,7 +25,8 @@ class IsTherePossibleMovesUseCase {
             //take a row line
             val row = boardGame[rowIndex]
             //if the first cell is the default value (there's empty cells) keep playing
-            if (row[0] == DEFAULT_VALUE) return MoveNumberResult(boardGame)
+            if (row[0] == DEFAULT_VALUE)
+                return playingBoardGame
 
             for (index in 0..<row.size - 1) {
                 val currentCell = row[index]
@@ -28,7 +34,8 @@ class IsTherePossibleMovesUseCase {
 
                 //if the next cell is the default value (there's empty cells)
                 //or is the same value as the current cell keep playing
-                if (nextCell == DEFAULT_VALUE || currentCell == nextCell) return MoveNumberResult(boardGame)
+                if (nextCell == DEFAULT_VALUE || currentCell == nextCell)
+                    return playingBoardGame
 
             }
         }
@@ -37,20 +44,23 @@ class IsTherePossibleMovesUseCase {
         for (columnIndex in 0..<boardSize) {
 
             val firstCell = boardGame[0][columnIndex]
-            if (firstCell == DEFAULT_VALUE) return MoveNumberResult(boardGame)
+            if (firstCell == DEFAULT_VALUE)
+                return playingBoardGame
 
             for (rowIndex in 0..<boardSize - 1) {
                 val currentCell = boardGame[rowIndex][columnIndex]
                 val nextCell = boardGame[rowIndex + 1][columnIndex]
 
-                if (nextCell == DEFAULT_VALUE || currentCell == nextCell) return MoveNumberResult(boardGame)
+                if (nextCell == DEFAULT_VALUE || currentCell == nextCell)
+                    return playingBoardGame
             }
         }
 
-        //At this point there's no possible next move so return the isGameOver property as true
+        //At this point there's no possible next move so returns GAME_OVER as the status
         return MoveNumberResult(
             boardGame = boardGame,
-            isGameOver = true
+            gameStatus = GAME_OVER,
+            numberToWin = numberToWin
         )
     }
 }
